Log Replay: Improved speed and position modifiers as well as cpu usage.
The Log replay thread now runs its own mavlink parser, which offloads
some of the CPU usage from the UI thread. This should make the UI a bit
more responsive when running the replay at higher speeds.

The position display has also been changed to be a user-settable slider,
for allowing the user to jump around in the log file. This will make
graphs look a bit odd if the user is constantly jumping, but that is to be
expected.